# feat(react-router): `params.parse` and `params.stringify` instead of `parseParams` and `stringifyParams`