… docs/framework/react/api/router/RouteOptionsType.md
… packages/react-router/src/index.tsx
… packages/react-router/tests/useLoaderData.test-d.tsx