@schiller-manuel 2024-06-25
LGTM

can you please:

- update documentation (deprecate old format, add new format)
- update examples to use the new format