@chorobin 2024-06-25
> LGTM
> 
> can you please:
> 
> * update documentation (deprecate old format, add new format)
> * update examples to use the new format

You missed the important bit... We need tests around all this 😂