@bpinto 2024-06-25
Sorry if I missed it, but is this compatible with declaring these on the routeR and not on each individual route?